agents/radar/pyproject.toml
[project]
name = "agent-lab-radar"
version = "0.0.1"
description = "Radar agent — discovery engine"
requires-python = ">=3.12"
dependencies = [
    "agent-lab-shared",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-community>=0.3",
    "httpx>=0.27",
    "pydantic>=2.9",
    "python-dotenv>=1.0",
    "beautifulsoup4>=4.12",
    "feedparser>=6.0",
    "click>=8.1",
    "langgraph>=0.4",
    "ag-ui-protocol>=0.1",
    "ag-ui-langgraph>=0.0.30",
    "copilotkit>=0.1.74",
]

[project.scripts]
radar-push = "radar.cron:main"
radar-serve = "radar.main:serve"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv.sources]
agent-lab-shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/radar"]
